+++ drivers/rtl8139.h
@@ -5,7 +5,7 @@
 
 // extern uint32_t RTL8139BaseAddress;
 
-uint32_t identifyRTL8139();
+uint8_t identifyRTL8139();
 uint8_t initializeRTL8139();
 
 #endif
